+++ src/Module/Api/Method/HandshakeMethod.php
@@ -64,7 +64,7 @@ final class HandshakeMethod
     public static function handshake(array $input): bool
     {
         $now_time   = time();
-        $timestamp  = preg_replace('/[^0-9]/', '', $input['timestamp'] ?? $now_time);
+        $timestamp  = (int)preg_replace('/[^0-9]/', '', $input['timestamp'] ?? $now_time);
         $passphrase = $input['auth'];
         if (empty($passphrase)) {
             $passphrase = Core::get_post('auth');
